fibo

+++ functions_nested_loops/102-fibonacci.c
@@ -1,27 +1,25 @@
 #include <stdio.h>
 #include "main.h"
 /**
- * main - Prints the first 50 Fibonacci numbers.
+ * main - printea los 50 fibo
  *
- * Description: This program generates and prints the first 50
- * Fibonacci numbers starting from 1 and 2, separated by a comma
- * and a space.
+ * Description: genera fibonacci 50 primeros 
  *
- * Return: Always returns 0 (success).
+ * Return: siempre exitoso 0
  */
 int main(void) {
-    long long int fib1 = 1, fib2 = 2;
-    long long int next_fib;
+    unsigned long fib1 = 1, fib2 = 2;
+    unsigned long next_fib;
     int count;
 
     for (count = 1; count <= 50; count++) {
         if (count == 1) {
-            printf("%lld", fib1);
+            printf("%lu", fib1);
         } else if (count == 2) {
-            printf(", %lld", fib2);
+            printf(", %lu", fib2);
         } else {
             next_fib = fib1 + fib2;
-            printf(", %lld", next_fib);
+            printf(", %lu", next_fib);
             fib1 = fib2;
             fib2 = next_fib;
         }
